// Game.Engine/wwwroot/src/game_generated.ts
export class Vec2 {
    bb: flatbuffers.ByteBuffer | null = null;

    bb_pos = 0;
    /**
     * @param number i
     * @param flatbuffers.ByteBuffer bb
     * @returns Vec2
     */
    __init(i: number, bb: flatbuffers.ByteBuffer): Vec2 {
        this.bb_pos = i;
        this.bb = bb;
        return this;
    }

    /**
     * @returns number
     */
    x(): number {
        return this.bb!.readInt16(this.bb_pos);
    }

    /**
     * @returns number
     */
    y(): number {
        return this.bb!.readInt16(this.bb_pos + 2);
    }

    /**
     * @param flatbuffers.Builder builder
     * @param number x
     * @param number y
     * @returns flatbuffers.Offset
     */
    static createVec2(builder: flatbuffers.Builder, x: number, y: number): flatbuffers.Offset {
        builder.prep(2, 4);
        builder.writeInt16(y);
        builder.writeInt16(x);
        return builder.offset();
    }
}

/**
 * @constructor
 */
export class NetBody {
    bb: flatbuffers.ByteBuffer | null = null;

    bb_pos = 0;
    /**
     * @param number i
     * @param flatbuffers.ByteBuffer bb
     * @returns NetBody
     */
    __init(i: number, bb: flatbuffers.ByteBuffer): NetBody {
        this.bb_pos = i;
        this.bb = bb;
        return this;
    }

    /**
     * @returns number
     */
    id(): number {
        return this.bb!.readUint32(this.bb_pos);
    }

    /**
     * @returns number
     */
    definitionTime(): number {
        return this.bb!.readUint32(this.bb_pos + 4);
    }

    /**
     * @param Vec2= obj
     * @returns Vec2|null
     */
    originalPosition(obj?: Vec2): Vec2 | null {
        return (obj || new Vec2()).__init(this.bb_pos + 8, this.bb!);
    }

    /**
     * @param Vec2= obj
     * @returns Vec2|null
     */
    velocity(obj?: Vec2): Vec2 | null {
        return (obj || new Vec2()).__init(this.bb_pos + 12, this.bb!);
    }

    /**
     * @returns number
     */
    originalAngle(): number {
        return this.bb!.readInt8(this.bb_pos + 16);
    }

    /**
     * @returns number
     */
    angularVelocity(): number {
        return this.bb!.readInt8(this.bb_pos + 17);
    }

    /**
     * @returns number
     */
    size(): number {
        return this.bb!.readUint8(this.bb_pos + 18);
    }

    /**
     * @returns number
     */
    sprite(): number {
        return this.bb!.readUint16(this.bb_pos + 20);
    }

    /**
     * @returns number
     */
    mode(): number {
        return this.bb!.readUint8(this.bb_pos + 22);
    }

    /**
     * @returns number
     */
    group(): number {
        return this.bb!.readUint32(this.bb_pos + 24);
    }

    /**
     * @param flatbuffers.Builder builder
     * @param number id
     * @param number definitionTime
     * @param number originalPosition_x
     * @param number originalPosition_y
     * @param number velocity_x
     * @param number velocity_y
     * @param number originalAngle
     * @param number angularVelocity
     * @param number size
     * @param number sprite
     * @param number mode
     * @param number group
     * @returns flatbuffers.Offset
     */
    static createNetBody(
        builder: flatbuffers.Builder,
        id: number,
        definitionTime: number,
        originalPosition_x: number,
        originalPosition_y: number,
        velocity_x: number,
        velocity_y: number,
        originalAngle: number,
        angularVelocity: number,
        size: number,
        sprite: number,
        mode: number,
        group: number
    ): flatbuffers.Offset {
        builder.prep(4, 28);
        builder.writeInt32(group);
        builder.pad(1);
        builder.writeInt8(mode);
        builder.writeInt16(sprite);
        builder.pad(1);
        builder.writeInt8(size);
        builder.writeInt8(angularVelocity);
        builder.writeInt8(originalAngle);
        builder.prep(2, 4);
        builder.writeInt16(velocity_y);
        builder.writeInt16(velocity_x);
        builder.prep(2, 4);
        builder.writeInt16(originalPosition_y);
        builder.writeInt16(originalPosition_x);
        builder.writeInt32(definitionTime);
        builder.writeInt32(id);
        return builder.offset();
    }
}